Docstrings

Public API

EvidentialDeepLearning.Dirichlet — Type

Dirichlet{N,T}

Statically sized Dirichlet distributions. In contrast to say Distributions.Dirichlet the the number of categories is encoded as a type parameter.

EvidentialDeepLearning.DirichletClassification — Type

DirichletClassification(Val(k))

Layer that takes a tensor of floats of size (k*n, b) and turns it into a tensor of Dirichlet distributions of size (n,b).

EvidentialDeepLearning.NIG — Type

NIG(μ, ν, α, β)

Normal-inverse-gamma distribution.

EvidentialDeepLearning.NIGRegression — Type

NormalInverseGamma regression layer. Takes a tensor of floats of size (4c, n) and outputs a tensor of NormalInverseGamma distributions of size (c,n).

EvidentialDeepLearning.posterior_predictive — Method

posterior_predictive(nig::NIG)

Return the posterior_predictive distribution, where the statistical model is the two parameter family Normal(μ,σ²) and the posterior is given by nig.

Internals

EvidentialDeepLearning.kl_uniform — Method

kl_uniform(d::Dirichlet)

Given P = Dirichlet(α) compute the Kullback Leibler divergence with respect to the uniform Dirichlet distribution:

`D_KL(P || Dirichlet(1,1,1...))`

EvidentialDeepLearning.std_aleatoric — Method

std_aleatoric(o) = √(var_aleatoric(o))

EvidentialDeepLearning.std_epistemic — Method

std_epistemic(o) = √(var_epistemic(o))

EvidentialDeepLearning.studentνμσ — Method

studentνμσ(ν,μ,σ)

Create a student t distributions with

ν degrees of freedom

μ location (= mean)

σ scale (≈ std deviation)

EvidentialDeepLearning.var_aleatoric — Function

var_aleatoric(d)

Alleatoric variance of predict(d). Also known as statistical uncertainty.

See also var_epistemic.

EvidentialDeepLearning.var_epistemic — Function

var_epistemic(d)

Epistemic variance of predict(d). Also known as systematic uncertainty.

See also var_aleatoric.